+++ activerecord/lib/active_record/associations/association.rb
@@ -235,7 +235,12 @@ module ActiveRecord
         # Returns true if inverse association on the given record needs to be set.
         # This method is redefined by subclasses.
         def invertible_for?(record)
-          inverse_reflection_for(record)
+          foreign_key_for?(record) && inverse_reflection_for(record)
+        end
+
+        # Returns true if record contains the foreign_key
+        def foreign_key_for?(record)
+          record.attributes.has_key? reflection.foreign_key
         end
 
         # This should be implemented to return the values of the relevant key(s) on the owner,